+++ packages/jest-haste-map/src/crawlers/__tests__/index.test.ts
@@ -30,6 +30,7 @@ const crawlerOptions = {
   ignore: () => false,
   rootDir: '/root',
   roots: ['/root'],
+  watchmanSockname: undefined,
 };
 
 const mockResult = {